WhatsAutomation/requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
cryptography==36.0.1
et-xmlfile==1.1.0
h11==0.12.0
idna==3.3
numpy==1.21.5
openpyxl==3.0.9
outcome==1.1.0
pandas==1.3.5
Pillow==8.4.0
pycparser==2.21
pyOpenSSL==21.0.0
python-dateutil==2.8.2
pytz==2021.3
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tk==0.1.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wsproto==1.0.0
